requirements.txt
numpy
gym
stable-baselines3
tkinter
Pillow
